+++ HISTORY.rst
@@ -3,6 +3,16 @@
 History
 -------
 
+1.1.0 (2014-12-XX)
+++++++++++++++++++
+
+* The pure Python reader now supports an optional file and memory mode in
+  addition to the existing memory-map mode. If your Python does not provide
+  the ``mmap`` module, the file mode will be used by default.
+* The preferred method for opening a database is now the ``open_database``
+  function in ``maxminddb``. This function now takes an optional file
+  ``mode``.
+
 1.0.0 (2014-09-22)
 ++++++++++++++++++